MCR-1657 use IntelliJ Code Cleanup to fix a bunch of codacy issues (#444)
fixed Java 5, 7 & 8 issues
fixed Log4J issues 
organized imports
reformatted code

+++ mycore-xeditor/src/main/java/org/mycore/frontend/xeditor/validation/MCRMaxDecimalValidator.java
@@ -50,7 +50,7 @@ public class MCRMaxDecimalValidator extends MCRDecimalValidator {
         if (d == null) {
             return false;
         } else {
-            return d.doubleValue() <= max;
+            return d <= max;
         }
     }
 }
