+++ mycore-xeditor/src/main/java/org/mycore/frontend/xeditor/validation/MCRMinDecimalValidator.java
@@ -50,7 +50,7 @@ public class MCRMinDecimalValidator extends MCRDecimalValidator {
         if (d == null) {
             return false;
         } else {
-            return min <= d.doubleValue();
+            return min <= d;
         }
     }
 }
